исправил замечания(создал реалмСервис и поменял принт на принт.локализейдДескрипшн)
