+++ VKapp/VKapp/Services/NetworkService/NetworkService.swift
@@ -77,43 +77,6 @@ final class NetworkService {
         return dataImage
     }
 
-    func saveFriendToRealm(_ friends: [Friend]) {
-        do {
-            let configuration = Realm.Configuration(deleteRealmIfMigrationNeeded: true)
-            let realm = try Realm(configuration: configuration)
-            try realm.write {
-                realm.add(friends, update: .modified)
-                print(realm.configuration.fileURL)
-            }
-        } catch {
-            print(error)
-        }
-    }
-
-    func saveGroupToRealm(_ groups: [Group]) {
-        do {
-            let realm = try Realm()
-            try realm.write {
-                realm.add(groups, update: .modified)
-                print(realm.configuration.fileURL)
-            }
-        } catch {
-            print(error)
-        }
-    }
-
-    func savePhotosToRealm(_ photos: [Photos]) {
-        do {
-            let realm = try Realm()
-            try realm.write {
-                realm.add(photos, update: .modified)
-                print(realm.configuration.fileURL)
-            }
-        } catch {
-            print(error)
-        }
-    }
-
     func createURLToLoadWebView() -> URL? {
         var urlComponents = URLComponents()
         urlComponents.scheme = Constants.scheme
